Added Hyprland Workspace + Submap support
Added a Hyprland IPC following widget to keep track of your current
workspaces and submap.
There is still more I can add to it, but I don't think I really need to.

Added a way to add more compositor specific widgets, but it isn't that
great and I should find another way.

That's basically it.