Updated widget error management and fixed flake.nix
Make all widget builders return a result, so that you can use the '?'
operator inside of the functions, and it will log it automatically.

I finally relized the issue with the flake, it was that buildInputs had
a capital B when it shouldn't have...

Rework of how the workspaces widget is selected, so that it defaults to
a warning that no workspace widget was included in the build, along with
also using results for the functions.

A cleanup will be needed in soon.