Logging Changes and other improvements.
Added more trace logging, and just made the logging a little better (for
the little of it there is)

Removed Regex dependency, because regex was really not needed, and I was
using it to check the start of a string...

Added back LTO and strip to the release profile, I forgot I turned it
off for testing earlier.

Likely some other stuff.